Introduce lint job in CI
This commit adds a lint job to the CI pipeline. It uses an up
to date Markdown linter https://github.com/DavidAnson/markdownlint-cli2.

I looked for an alternative from the Python ecosystem (to stay
more consistent with this project), but the only result was an
outdated and unmaintained project.

It would be possible to add a `make format` command as was
suggested in the original GitHub issue, but this would require
Node.js to be available in the dev environment, so I thought
setting up CI only would be good enough for now.

Note: it should be as easy as adding the following to the
Makefile:
```
format:
	npx markdownlint-cli2-fix "**/*.md"
```

Closes https://github.com/openfisca/openfisca-doc/issues/267